Add files via upload
This is RAG application that uses PDF provided the user , and allows the user to interact with PDF . Initially the user has to submit their Groq API key to initiate the conversational chatbot. The Chatbot uses the PDF uploaded by the user . The PDF taken as an input is converted into vector document using HuggingFaceEmbeddings and further stored into a vector database , FAISS . Further the vector store database is converted into a retriever for better use of the content search. 

The chatbot also takes the conversation history into consideration while interacting with the user . Here the system prompt for the AI model is : 
                "You are an assistant for question-answering tasks. "
                "Use the following pieces of retrieved context to answer "
                "the question. If you don't know the answer, say that you "
                "don't know. Use three sentences maximum and keep the "
                "answer concise."
Due to the above prompt the chatbot answers the question more efficiently and with better accuracy.

For answer the questions asked the chatbot uses a chain : 
       question_answer_chain=create_stuff_documents_chain(llm,qa_prompt)
        rag_chain=create_retrieval_chain(history_aware_retriever,question_answer_chain)

With the help of this chain the LLM model is invoke with the PDF and chat history as context information . Here the LLM model used is : "Gemma2-9b-It" and the model is used with the help of AI inferencing technique provided by Groq .

Here the model uses the entire context and with its knowledge it answers the user provided questions. 

The platform used to make the frontend is streamlit.

Important Links :

Langchain:  https://python.langchain.com/v0.2/docs/integrations/components/
Groq : https://groq.com/about-us/
Hugging Face : https://huggingface.co/
Streamlit : https://streamlit.io/